backend/api/requirements.txt
Flask>=2.0
werkzeug>=2.0,<2.2
flask-ngrok==0.0.25
google-generativeai
langchain ==0.0.2
flask-cors==3.0.10
